pyproject.toml
[project]
name = "ETL"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "kozka.c",email = "corentin.kozk@gmail.com"},
    {name = "vu.f",email = "frederick.vu.pro@gmail.com"},
    {name = "ibrahim.y",email = "youva.ibrahim@gmail.com"},

]
dependencies = [
    "pandas>=2.2.0",
    "pyarrow>=15.0.0",
    "pyyaml>=6.0.1",
    "termcolor>=2.4.0",
    "argcomplete>=3.2.2",
    "requests>=2.31.0",
    "python-dotenv>=1.0.1",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "CC-BY-NC-ND"}


[tool.pdm]
distribution = false
